requirements.txt
fastapi 
uvicorn[standard] 
biopython 
numpy 
pandas 
